requirements.txt
numpy==1.18.5
Flask==1.1.2
tensorflow==2.2.0
Pillow==7.1.2
scikit-image==0.17.2
Werkzeug==1.0.1
h5py==2.10.0
gunicorn==20.0.4
gevent==20.6.2
